See #24. Treat Material with transparent color of 0 as non-dielectric
This fixes an issue where sample reflective scenes would render black
objects because they not only were reflective but also had the
refractionIndex field set to non-zero and transparent color was 0.
Now, if this is the case, it will treat it still as if the material was
not transparent and only reflective, allowing tests to pass.